reduce scope to private location only
