+++ .github/workflows/test.yml
@@ -102,9 +102,6 @@ jobs:
           - '8.11.4'
           - '8.12.2'
           - '8.13.4'
-          - '8.14.0'
-          - '8.14.1'
-          - '8.14.2'
           - '8.14.3'
     steps:
       - uses: actions/checkout@v4
